Created boundary classes. General refactor
