+++ src/main/java/it/uniroma2/alessandrolioi/jira/exceptions/JiraRestException.java
@@ -2,11 +2,11 @@ package it.uniroma2.alessandrolioi.jira.exceptions;
 
 import java.io.Serial;
 
-public class JiraRESTException extends Exception {
+public class JiraRestException extends Exception {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    public JiraRESTException(String reason, Throwable cause) {
+    public JiraRestException(String reason, Throwable cause) {
         super(reason, cause);
     }
 }
